Refactored packages according to config
