+++ src/main/java/org/radarbase/hdfs/Application.kt
@@ -22,14 +22,15 @@ import org.apache.hadoop.fs.Path
 import org.radarbase.hdfs.accounting.Accountant
 import org.radarbase.hdfs.accounting.HdfsRemoteLockManager
 import org.radarbase.hdfs.accounting.RemoteLockManager
+import org.radarbase.hdfs.config.CommandLineArgs
 import org.radarbase.hdfs.config.RestructureConfig
-import org.radarbase.hdfs.data.Compression
-import org.radarbase.hdfs.data.FileCacheStore
-import org.radarbase.hdfs.data.RecordConverterFactory
-import org.radarbase.hdfs.data.StorageDriver
+import org.radarbase.hdfs.compression.Compression
+import org.radarbase.hdfs.format.RecordConverterFactory
+import org.radarbase.hdfs.path.RecordPathFactory
+import org.radarbase.hdfs.storage.StorageDriver
 import org.radarbase.hdfs.util.ProgressBar.Companion.format
 import org.radarbase.hdfs.util.Timer
-import org.radarbase.hdfs.util.commandline.CommandLineArgs
+import org.radarbase.hdfs.worker.FileCacheStore
 import org.slf4j.LoggerFactory
 import java.io.IOException
 import java.nio.file.Files
