+++ src/main/java/org/radarbase/hdfs/config/NonEmptyValidator.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 The Hyve
+ * Copyright 2018 The Hyve
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package org.radarbase.hdfs.data
+package org.radarbase.hdfs.config
 
-import java.io.Closeable
-import java.io.Flushable
-import java.io.IOException
-import org.apache.avro.generic.GenericRecord
+import com.beust.jcommander.IParameterValidator
+import com.beust.jcommander.ParameterException
 
-/** Converts a GenericRecord to Java primitives or writes it to file.  */
-interface RecordConverter : Flushable, Closeable {
-    @Throws(IOException::class)
-    fun writeRecord(record: GenericRecord): Boolean
-
-    fun convertRecord(record: GenericRecord): Map<String, Any?>
+class NonEmptyValidator : IParameterValidator {
+    override fun validate(name: String, value: String?) {
+        if (value?.isEmpty() == true) {
+            throw ParameterException("Parameter " + name + " should be supplied. "
+                    + "It cannot be empty or null. (found " + value + ")."
+                    + "Please run with --help or -h for more information.")
+        }
+    }
 }
